Move stake pools to constants file
